Cast action to proper type (unsigned)

+++ 3RVX/HotkeyInfo.cpp
@@ -49,7 +49,7 @@ double HotkeyInfo::ArgToDouble(int argIdx) {
 std::wstring HotkeyInfo::ToString() {
     std::wstring combination = HotkeyManager::HotkeysToString(keyCombination);
     std::wstring act = L"(none)";
-    if (action > 0 && action < ActionNames.size()) {
+    if (action > 0 && (unsigned int) action < ActionNames.size()) {
         act = ActionNames[action];
     }
     std::wstring argStrs;
